RetroPlayer: Fix compiler warnings on Windows (use size_t for memory sizes)

+++ xbmc/cores/RetroPlayer/RetroPlayerAudio.cpp
@@ -112,14 +112,14 @@ bool CRetroPlayerAudio::OpenEncodedStream(AVCodecID codec, unsigned int samplera
   return true; //! @todo
 }
 
-void CRetroPlayerAudio::AddData(const uint8_t* data, unsigned int size)
+void CRetroPlayerAudio::AddData(const uint8_t* data, size_t size)
 {
   if (m_bAudioEnabled)
   {
     if (m_pAudioStream)
     {
-      const unsigned int frameSize = m_pAudioStream->GetChannelCount() * (CAEUtil::DataFormatToBits(m_pAudioStream->GetDataFormat()) >> 3);
-      m_pAudioStream->AddData(&data, 0, size / frameSize);
+      const size_t frameSize = m_pAudioStream->GetChannelCount() * (CAEUtil::DataFormatToBits(m_pAudioStream->GetDataFormat()) >> 3);
+      m_pAudioStream->AddData(&data, 0, static_cast<unsigned int>(size / frameSize));
     }
   }
 }
