+++ xbmc/cores/RetroPlayer/RetroPlayerAudio.h
@@ -41,7 +41,7 @@ namespace RETRO
     unsigned int NormalizeSamplerate(unsigned int samplerate) const override;
     bool OpenPCMStream(AEDataFormat format, unsigned int samplerate, const CAEChannelInfo& channelLayout) override;
     bool OpenEncodedStream(AVCodecID codec, unsigned int samplerate, const CAEChannelInfo& channelLayout) override;
-    void AddData(const uint8_t* data, unsigned int size) override;
+    void AddData(const uint8_t* data, size_t size) override;
     void CloseStream() override;
 
     void Enable(bool bEnabled) { m_bAudioEnabled = bEnabled; }
